pyproject.toml
[project]
name = "chewy-notification"
version = "0.1.0"
description = "A universal Django notification system supporting multiple channels (Bark, Ntfy, Email, Feishu)"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Chewy Team", email = "contact@example.com"},
]
keywords = ["django", "notification", "push", "email", "bark", "ntfy", "feishu"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "django-filter>=23.0",
    "requests>=2.31",
]

[project.optional-dependencies]
celery = [
    "celery>=5.3",
    "redis>=5.0",
]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "black>=23.0",
    "flake8>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chewy_notification"]
